% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauchodown.R
\docType{package}
\name{gauchodown}
\alias{gauchodown}
\alias{gauchodown-package}
\title{gauchodown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
gauchodown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown
}
