% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsfdown.R
\docType{package}
\name{ucsfdown}
\alias{ucsfdown}
\alias{ucsfdown-package}
\title{ucsfdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
ucsfdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown
}
