% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggiedown.R
\docType{package}
\name{aggiedown}
\alias{aggiedown}
\alias{aggiedown-package}
\title{gauchodown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
gauchodown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown
}
